Add test service to main package
